Reworked price field
